remove double test with only a single change
